+++ pricing/src/epilot/models/shared/availabilityfilters.py
@@ -10,7 +10,6 @@ from typing import Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
-
 @dataclasses.dataclass
 class AvailabilityFilters:
     r"""Availability filters dimensions"""