+++ pricing/src/epilot/models/shared/catalogsearch.py
@@ -9,7 +9,6 @@ from typing import Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
-
 @dataclasses.dataclass
 class CatalogSearch:
     r"""A catalog search payload"""