+++ pricing/src/epilot/models/shared/checkoutcartresult.py
@@ -2,17 +2,17 @@
 
 from __future__ import annotations
 import dataclasses
+from ..shared import order as shared_order
 from dataclasses_json import Undefined, dataclass_json
 from epilot import utils
-from typing import Any, Optional
+from typing import Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
-
 @dataclasses.dataclass
 class CheckoutCartResult:
     r"""The cart checkout result"""
-    order: Optional[dict[str, Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('order'), 'exclude': lambda f: f is None }})
+    order: Optional[shared_order.Order] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('order'), 'exclude': lambda f: f is None }})
     r"""The order entity"""