+++ pricing/src/epilot/models/shared/ordersource.py
@@ -8,7 +8,6 @@ from typing import Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
-
 @dataclasses.dataclass
 class OrderSource:
     r"""The order generation source"""